python/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
ariadne==0.22.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-multipart==0.0.6
aiodataloader==0.4.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-json-logger==2.0.7
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0

